+++ rest/src/main/kotlin/com/gitlab/kordlib/rest/builder/role/RoleCreateBuilder.kt
@@ -2,7 +2,7 @@ package com.gitlab.kordlib.rest.builder.role
 
 import com.gitlab.kordlib.common.entity.Permissions
 import com.gitlab.kordlib.rest.builder.AuditRequestBuilder
-import com.gitlab.kordlib.rest.builder.KordDsl
+import com.gitlab.kordlib.common.annotation.KordDsl
 import com.gitlab.kordlib.rest.json.request.GuildRoleCreateRequest
 import java.awt.Color