+++ rest/src/main/kotlin/com/gitlab/kordlib/rest/builder/user/GroupDMCreateBuilder.kt
@@ -1,7 +1,7 @@
 package com.gitlab.kordlib.rest.builder.user
 
 import com.gitlab.kordlib.common.entity.Snowflake
-import com.gitlab.kordlib.rest.builder.KordDsl
+import com.gitlab.kordlib.common.annotation.KordDsl
 import com.gitlab.kordlib.rest.builder.RequestBuilder
 import com.gitlab.kordlib.rest.json.request.GroupDMCreateRequest